Checkpoint: moved vni_provider.py under core/
